graph class
two public function in the graph class
one is to find the a shortest path from one specific node to another
specific node.
another one is to find the shortest path from a specific node and go
through all of the nodes.
Both function will return the distance of the path. I am not sure how to
return the path, so far I just cout it to the screen.